+++ lib/button/Button.h
@@ -1,15 +1,14 @@
 #include "Bounce2.h"
-#include "Arduino.h"
 
 class Button : public Bounce
 {
     public:
         Button ();
-        Button (uint8_t pin);
-        Button (uint8_t pin, unsigned int debounceDelay);
+        Button (byte pin);
+        Button (byte pin, byte debounceDelay);
 
-        void init(uint8_t pin, unsigned int debounceDelay);
+        void init(byte pin, byte debounceDelay);
         bool isPressed();
         unsigned long getStateTime();
-        unsigned int getSpeedLevel();
+        byte getSpeedLevel();
 };
