+++ lib/stakles/Stakles.h
@@ -15,28 +15,28 @@ class Stakles
     void process();
 
     // POWER
-    uint8_t power_port;
-    void power_init(uint8_t pin);
+    byte power_port;
+    void power_init(byte pin);
     bool is_power_on();
 
     // LED
-    void led_init(uint8_t CLK, uint8_t DIO);
+    void led_init(byte CLK, byte DIO);
 
     // BUTTON
-    void button_L_init(uint8_t pin);
-    void button_R_init(uint8_t pin);
+    void button_L_init(byte pin);
+    void button_R_init(byte pin);
 
     // ENCODER
-    void encoder_init(uint8_t l_pin, uint8_t r_pin);
+    void encoder_init(byte l_pin, byte r_pin);
 
     // MOVE
     void move_to_l ();
     void move_to_r ();
-    void set_height  (uint32_t height);
+    void set_height  (unsigned int height);
     float get_height();
 
   private:
-    unsigned long int height = 320; // 40mm 
+    unsigned int height = 320; // 40mm 
 
     // DEBUG
     unsigned long debug_time = 0;
@@ -44,7 +44,7 @@ class Stakles
     unsigned long count_x = 0;
 
     // POWER
-    uint8_t power_pin;
+    byte power_pin;
 
     // CONFIG
     Config config;
@@ -68,8 +68,8 @@ class Stakles
     bool button_update();
 
     // MOVE
-    const float        delta_360 = 0.2; //mm
-    const unsigned int steps_per_round = 16; // 8 skyles * 2 sensoriai  
-    const float        delta_step = this->delta_360 / this->steps_per_round;
+    const float delta_360 = 0.2; //mm
+    const byte  steps_per_round = 16; // 8 skyles * 2 sensoriai  
+    const float delta_step = this->delta_360 / this->steps_per_round;
 
 };
